added Item to fk lookup & test in ItemType

+++ src/main/java/com/mcommings/campaigner/services/items/ItemTypeService.java
@@ -3,24 +3,31 @@ package com.mcommings.campaigner.services.items;
 import com.mcommings.campaigner.interfaces.items.IItemType;
 import com.mcommings.campaigner.models.RepositoryHelper;
 import com.mcommings.campaigner.models.items.ItemType;
+import com.mcommings.campaigner.repositories.items.IItemRepository;
 import com.mcommings.campaigner.repositories.items.IItemTypeRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import static com.mcommings.campaigner.enums.ErrorMessage.*;
+import static com.mcommings.campaigner.enums.ForeignKey.FK_ITEM_TYPE;
 
 @Service
 public class ItemTypeService implements IItemType {
 
     private final IItemTypeRepository itemTypeRepository;
+    private final IItemRepository itemRepository;
 
     @Autowired
-    public ItemTypeService(IItemTypeRepository itemTypeRepository) {
+    public ItemTypeService(IItemTypeRepository itemTypeRepository, IItemRepository itemRepository) {
         this.itemTypeRepository = itemTypeRepository;
+        this.itemRepository = itemRepository;
     }
 
     @Override
@@ -47,10 +54,9 @@ public class ItemTypeService implements IItemType {
         if (RepositoryHelper.cannotFindId(itemTypeRepository, itemTypeId)) {
             throw new IllegalArgumentException(DELETE_NOT_FOUND.message);
         }
-// TODO: uncomment when class that uses ItemType as a fk is added
-//        if (RepositoryHelper.isForeignKey(getReposWhereItemTypeIsAForeignKey(), FK_ITEM_TYPE.columnName, itemTypeId)) {
-//            throw new DataIntegrityViolationException(DELETE_FOREIGN_KEY.message);
-//        }
+        if (RepositoryHelper.isForeignKey(getReposWhereItemTypeIsAForeignKey(), FK_ITEM_TYPE.columnName, itemTypeId)) {
+            throw new DataIntegrityViolationException(DELETE_FOREIGN_KEY.message);
+        }
 
         itemTypeRepository.deleteById(itemTypeId);
     }
@@ -66,8 +72,7 @@ public class ItemTypeService implements IItemType {
         itemTypeToUpdate.setDescription(itemType.getDescription());
     }
 
-// TODO: uncomment when class that uses ItemType as a fk is added
-//    private List<CrudRepository> getReposWhereItemTypeIsAForeignKey() {
-//        return new ArrayList<>(Arrays.asList());
-//    }
+    private List<CrudRepository> getReposWhereItemTypeIsAForeignKey() {
+        return new ArrayList<>(Arrays.asList(itemRepository));
+    }
 }
